server/mqtt_bridge/requirements.txt
fastapi
uvicorn
paho-mqtt
kafka-python
python-dotenv
